apps/api/pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "eu-comply-api"
version = "0.1.0"
description = "FastAPI control plane for EU-Comply."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.21.0",
  "alembic>=1.16.0",
  "asyncpg>=0.30.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.12",
  "greenlet>=3.2.0",
  "httpx>=0.28.1",
  "orjson>=3.10.18",
  "opentelemetry-api>=1.32.1",
  "opentelemetry-instrumentation-fastapi>=0.53b1",
  "opentelemetry-instrumentation-httpx>=0.53b1",
  "opentelemetry-instrumentation-sqlalchemy>=0.53b1",
  "opentelemetry-sdk>=1.32.1",
  "passlib[bcrypt]>=1.7.4",
  "pydantic>=2.11.3",
  "pydantic-settings>=2.8.1",
  "python-jose[cryptography]>=3.4.0",
  "redis[hiredis]>=6.0.0",
  "sqlalchemy>=2.0.40",
  "structlog>=25.3.0",
  "tenacity>=9.1.2",
  "uvicorn[standard]>=0.34.2"
]

[project.optional-dependencies]
dev = [
  "anyio>=4.9.0",
  "mypy>=1.15.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.5"
]

[tool.hatch.build.targets.wheel]
packages = ["src/eu_comply_api"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501"]
"src/eu_comply_api/api/deps.py" = ["B008"]
"src/eu_comply_api/api/routes/*.py" = ["B008"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
